Copy config JSON
config.json and generation_config.json change depending on your version of the transformers package. This change copies the files directly instead of applying the a byte delta, which might break depending on your version of the transformers package.

I also ran `format.sh` which lints the changes in the PR