Use `model id` + latest version
This updates the code to use `model_id` instead of `model_version_id`,
this way the latest, released version will always be used. This is
potentially less stable, but means we don't have to update the version
ID with each model release.

I also added a little code to handle the case where greedy sampling is
used to prevent the error that's raised by vLLM.